requirements.txt
Django==1.9.4
argparse==1.2.1
django-tastypie==0.13.3
mongoengine==0.10.6
pymongo==3.2.2
python-dateutil==2.5.2
python-mimeparse==1.5.1
six==1.10.0
wsgiref==0.1.2
